+++ modules/telegram/src/test/java/io/starter/tests/BaseTest.java
@@ -1,4 +1,4 @@
-package io.starter.units.handler;
+package io.starter.tests;
 
 import net.datafaker.Faker;
